Refactor server stop/start and generic requesting out to base class
